+++ lib/mix/tasks/install_runtime.ex
@@ -43,7 +43,7 @@ defmodule Mix.Tasks.Lightning.InstallRuntime do
 
   def packages do
     ~W(
-      @openfn/cli@1.3.2
+      @openfn/cli@1.8.1
       @openfn/language-common@latest
     )
   end